Разбил на компоненты и настроил окружение
