+++ public/index.html
@@ -34,6 +34,6 @@
     <p>130 291 movies inside</p>
   </section>
 </footer>
-  <script src="../src/main.js"></script>
+  <script src="./bundle.js"></script>
 </body>
 </html>
